improve interfact

+++ zoning/search/__main__.py
@@ -1,3 +1,4 @@
+import os
 import json
 
 import hydra
@@ -9,20 +10,20 @@ from zoning.utils import process
 
 
 def preprocess_search_target(town_file, district_file, eval_terms, output_file):
-    town = json.load(open(town_file))
-    districts = json.load(open(district_file))
-    search_targets = []
-    for district in districts:
-        if district.split("__")[0] not in town:
-            continue
-        town, district_short_name, district_full_name = district.split("__")
-        for term in eval_terms:
-            search_targets.append(
-                f"{term}__{town}__{district_short_name}__{district_full_name}"
-            )
-    with open(output_file, "w") as f:
-        json.dump(search_targets, f, indent=4)
-    return search_targets
+    if not os.path.exists(output_file):
+        town = json.load(open(town_file))
+        districts = json.load(open(district_file))
+        search_targets = []
+        for district in districts:
+            if district.split("__")[0] not in town:
+                continue
+            town, district_short_name, district_full_name = district.split("__")
+            for term in eval_terms:
+                search_targets.append(
+                    f"{term}__{town}__{district_short_name}__{district_full_name}"
+                )
+        with open(output_file, "w") as f:
+            json.dump(search_targets, f, indent=4)
 
 
 @hydra.main(version_base=None, config_path="../../config", config_name="base")
@@ -57,6 +58,7 @@ def main(config: ZoningConfig):
             raise ValueError(f"Search method {search_config.method} is not supported")
 
     # Construct the entire search query with all possible eval terms
+    
     preprocess_search_target(
         global_config.target_town_file,
         global_config.target_district_file,
